noteActions の実装
