--- src/hooks/Note/useUpdateNote/index.ts
@@ -1,1 +1,0 @@
-export { useUpdateNote } from './useUpdateNote';
